Move updater into amplipi python folder structure
If we made this into a seperate package it would require a seperate repo. This does not make sense at this point. Given this decision it will be hard to seperate the updater in the future. This would need to happen on a major release.
