+++ amplipi/updater/static/js/demo-config.js
