+++ amplipi/updater/static/js/demo-ui.js
